gulp: split tasks across multiple files
