+++ generator/component/temp.html
@@ -1,3 +1,5 @@
+<style>@import 'components/<%= name %>/<%= name %>.css'</style>
+
 <div class="<%= name %>">
   <h1>{{ vm.myName }}</h1>
 </div>
