+++ package.json
@@ -55,6 +55,7 @@
     "mocha": "^2.2.5",
     "open": "0.0.5",
     "phantomjs": "^1.9.17",
+    "require-dir": "^0.3.0",
     "run-sequence": "^1.1.0",
     "semver": "^4.3.6",
     "serve-static": "^1.9.2",
